Remove dead food seeker code
