+++ client/src/components/FoodSeeker/SearchResults/ResultsMap/MarkerHelpers.js
