+++ client/src/components/FoodSeeker/SearchResults/ResultsMap/styles.js
