+++ client/src/components/FoodSeeker/SearchResults/StakeholderPreview/StakeholderPreview.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import moment from "moment";
 import * as momentTz from "moment-timezone";
 import { makeStyles } from "@material-ui/core/styles";
-import { Button } from "../../../../components/UI";
+import { Button } from "../../../UI";
 import { useSiteContext } from "contexts/siteContext";
 
 import {
